fix(std): update last_index_byte function for GLIBC compatibility
Modify the last_index_byte function to ensure compatibility with GLIBC by adding a conditional compilation directive. This change prevents potential issues when the __neverc__ macro is defined. Additionally, update the test case to use long long instead of int64_t for better type consistency in integer parsing.

+++ tests/neverc/std/test_dot_syntax.c
@@ -100,7 +100,7 @@ static void test_strconv_dot_syntax(void) {
     n = strconv.format_int(-7, 10, buf, sizeof(buf));
     CHECK("strconv.format_int_neg", n > 0 && strcmp(buf, "-7") == 0);
 
-    int64_t val;
+    long long val;
     int ok = strconv.parse_int("12345", 10, &val);
     CHECK("strconv.parse_int_ok", ok == 0 && val == 12345);
